fix: make tests compile with forge-std@1.6.0
There was a breaking change introduced in `forge-std` at
https://github.com/foundry-rs/forge-std/pull/407 which breaks
compilation of `Rln.t.sol` with `forge-std@v1.6.0`.

This commit updates the dependency to v1.6.0 and adjusts the test source
such that it successfully compiles.

Another way to go about this would've been to just stick with `v1.5.6.`
and ensuring installation of that version.
However, I've decided to update the dependency to the latest stable
version instead.